chore(suite-common): feeUnitUtils
